doit: Split out cmd_sfc
